Downgrade gonja lib to v2.3.1 as the new fix for windows causing bugs
The bug is strange.

If I add my config file and process everything seems to look - except

If within the for loop, add a if block like the current test1.j2 case it
failed with index out of range in exec renderer.

So Downgrade for now and forget abt windows